+++ common/src/main/java/me/rufia/fightorflight/client/model/PokemonSpikeModel.java
@@ -18,7 +18,7 @@ import net.minecraft.world.entity.Entity;
 
 public class PokemonSpikeModel<T extends Entity> extends EntityModel<T> {
     // This layer location should be baked with EntityRendererProvider.Context in the entity renderer and passed into this model's constructor
-    public static final ModelLayerLocation LAYER_LOCATION = new ModelLayerLocation(ResourceLocation.fromNamespaceAndPath(CobblemonFightOrFlight.MODID, "pokemonspike"), "main");
+    public static final ModelLayerLocation LAYER_LOCATION = new ModelLayerLocation(ResourceLocation.fromNamespaceAndPath(CobblemonFightOrFlight.MODID, "spike"), "main");
     private final ModelPart Main;
 
     public PokemonSpikeModel(ModelPart root) {
@@ -31,22 +31,23 @@ public class PokemonSpikeModel<T extends Entity> extends EntityModel<T> {
 
         PartDefinition Main = partdefinition.addOrReplaceChild("Main", CubeListBuilder.create().texOffs(0, 0).addBox(-2.0F, -2.0F, -2.0F, 4.0F, 4.0F, 4.0F, new CubeDeformation(0.0F))
                 .texOffs(0, 8).addBox(-1.0F, -8.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F))
-                .texOffs(0, 8).addBox(-1.0F, 2.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offset(0.0F, 16.0F, 0.0F));
+                .texOffs(0, 8).addBox(-1.0F, 2.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offset(0.0F, 24.0F, 0.0F));
 
-        PartDefinition East_r1 = Main.addOrReplaceChild("East_r1", CubeListBuilder.create().texOffs(0, 8).addBox(-1.0F, -6.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offsetAndRotation(-8.0F, 0.0F, 0.0F, 0.0F, 0.0F, 1.5708F));
+        PartDefinition East_r1 = Main.addOrReplaceChild("East_r1", CubeListBuilder.create().texOffs(0, 8).addBox(-1.0F, -3.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offsetAndRotation(-5.0F, 0.0F, 0.0F, 0.0F, 0.0F, 1.5708F));
 
-        PartDefinition West_r1 = Main.addOrReplaceChild("West_r1", CubeListBuilder.create().texOffs(0, 8).addBox(-1.0F, -6.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offsetAndRotation(8.0F, 0.0F, 0.0F, 0.0F, 0.0F, -1.5708F));
+        PartDefinition West_r1 = Main.addOrReplaceChild("West_r1", CubeListBuilder.create().texOffs(0, 8).addBox(-1.0F, -3.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offsetAndRotation(5.0F, 0.0F, 0.0F, 0.0F, 0.0F, -1.5708F));
 
-        PartDefinition South_r1 = Main.addOrReplaceChild("South_r1", CubeListBuilder.create().texOffs(0, 8).addBox(-1.0F, -6.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offsetAndRotation(0.0F, 0.0F, 2.0F, -1.5708F, 0.0F, 0.0F));
+        PartDefinition South_r1 = Main.addOrReplaceChild("South_r1", CubeListBuilder.create().texOffs(0, 8).addBox(-1.0F, -3.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offsetAndRotation(0.0F, 0.0F, 5.0F, -1.5708F, 0.0F, 0.0F));
 
-        PartDefinition North_r1 = Main.addOrReplaceChild("North_r1", CubeListBuilder.create().texOffs(0, 8).addBox(-1.0F, -6.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offsetAndRotation(0.0F, 0.0F, -2.0F, 1.5708F, 0.0F, 0.0F));
+        PartDefinition North_r1 = Main.addOrReplaceChild("North_r1", CubeListBuilder.create().texOffs(0, 8).addBox(-1.0F, -3.0F, -1.0F, 2.0F, 6.0F, 2.0F, new CubeDeformation(0.0F)), PartPose.offsetAndRotation(0.0F, 0.0F, -5.0F, 1.5708F, 0.0F, 0.0F));
 
         return LayerDefinition.create(meshdefinition, 16, 16);
     }
 
     @Override
     public void setupAnim(Entity entity, float limbSwing, float limbSwingAmount, float ageInTicks, float netHeadYaw, float headPitch) {
-
+        this.Main.yRot = netHeadYaw * 0.017453292F;
+        this.Main.xRot = headPitch * 0.017453292F;
     }
 
     @Override