+++ common/src/main/java/me/rufia/fightorflight/client/renderer/PokemonSpikeRenderer.java
@@ -29,29 +29,31 @@ public class PokemonSpikeRenderer extends EntityRenderer<PokemonSpike> {
     }
 
     @Override
-    public void render(PokemonSpike entity, float entityYaw, float partialTicks, PoseStack poseStack, MultiBufferSource buffer, int packedLight){
+    public void render(PokemonSpike entity, float entityYaw, float partialTicks, PoseStack poseStack, MultiBufferSource buffer, int packedLight) {
         Color color = getColor(entity);
+        int inGroundTick = entity.getInGroundTick();
         poseStack.pushPose();
         float f = Mth.rotLerp(partialTicks, entity.yRotO, entity.getYRot());
         float g = Mth.lerp(partialTicks, entity.xRotO, entity.getXRot());
-        float h = (float) entity.tickCount + partialTicks;
-        poseStack.translate(0.0F, 0.15F, 0.0F);
+        float h = inGroundTick == -1 ? (float) entity.tickCount + partialTicks : inGroundTick;
+        poseStack.scale(0.75f, 0.75f, 0.75f);
+        //CobblemonFightOrFlight.LOGGER.info(Integer.toString(inGroundTick));
         poseStack.mulPose(Axis.YP.rotationDegrees(Mth.sin(h * 0.1F) * 180.0F));
         poseStack.mulPose(Axis.XP.rotationDegrees(Mth.cos(h * 0.1F) * 180.0F));
         poseStack.mulPose(Axis.ZP.rotationDegrees(Mth.sin(h * 0.15F) * 360.0F));
-        poseStack.scale(-0.5F, -0.5F, 0.5F);
+        poseStack.translate(0f, -1.5f, 0f);
         this.model.setupAnim(entity, 0.0F, 0.0F, 0.0F, f, g);
         VertexConsumer vertexConsumer = buffer.getBuffer(this.model.renderType(TEXTURE_LOCATION));
-        this.model.renderToBuffer(poseStack, vertexConsumer, packedLight, OverlayTexture.NO_OVERLAY);
-        poseStack.scale(1.5F, 1.5F, 1.5F);
-        VertexConsumer vertexConsumer2 = buffer.getBuffer(RENDER_TYPE);
         int colorCode = FastColor.ARGB32.colorFromFloat((float) color.getRed() / 255, (float) color.getGreen() / 255, (float) color.getBlue() / 255, 0.75F);
-        this.model.renderToBuffer(poseStack, vertexConsumer2, packedLight, OverlayTexture.NO_OVERLAY, colorCode);
+        this.model.renderToBuffer(poseStack, vertexConsumer, packedLight, OverlayTexture.NO_OVERLAY, colorCode);
         poseStack.popPose();
         super.render(entity, entityYaw, partialTicks, poseStack, buffer, packedLight);
     }
 
-    public Color getColor(PokemonSpike entity){
+    public Color getColor(PokemonSpike entity) {
+        if (entity.getElementalType() != null) {
+            return PokemonAttackEffect.getColorFromType(entity.getElementalType());
+        }
         return Color.white;
     }