+++ common/src/main/java/me/rufia/fightorflight/data/movedata/movedatas/MiscMoveData.java
@@ -1,9 +1,11 @@
 package me.rufia.fightorflight.data.movedata.movedatas;
 
+import com.cobblemon.mod.common.api.moves.Move;
 import com.cobblemon.mod.common.entity.pokemon.PokemonEntity;
 import me.rufia.fightorflight.PokemonInterface;
 import me.rufia.fightorflight.data.movedata.MoveData;
 import me.rufia.fightorflight.entity.PokemonAttackEffect;
+import me.rufia.fightorflight.utils.PokemonUtils;
 import net.minecraft.world.entity.LivingEntity;
 
 import java.util.Objects;
@@ -27,6 +29,10 @@ public class MiscMoveData extends MoveData {
         }
         if (Objects.equals("heal_per_50", getName())) {
             finalTarget.heal(0.5f * finalTarget.getMaxHealth());
+        }
+        if (Objects.equals("spikes", getName())) {
+            Move move = PokemonUtils.getStatusMove(pokemonEntity);
+            PokemonAttackEffect.spreadSpikes(pokemonEntity, move);
         }
     }
 }