+++ common/src/main/java/me/rufia/fightorflight/goals/targeting/PokemonProactiveTargetGoal.java
@@ -17,11 +17,7 @@ import java.util.function.Predicate;
 
 public class PokemonProactiveTargetGoal<T extends LivingEntity> extends NearestAttackableTargetGoal<T> {
 
-    public float safeDistanceSqr = 36;
-    //public PokemonProactiveTargetGoal(Mob mob, Class<T> targetType, float safeDistanceSqr, boolean mustSee, boolean mustReach) {
-    //    super(mob, targetType, mustSee, mustReach);
-    //    this.safeDistanceSqr = safeDistanceSqr;
-    //}
+    public float safeDistanceSqr;
 
     public PokemonProactiveTargetGoal(Mob mob, Class<T> targetType, float safeDistanceSqr, int randomInterval, boolean mustSee, boolean mustReach, @Nullable Predicate<LivingEntity> targetPredicate) {
         super(mob, targetType, randomInterval, mustSee, mustReach, targetPredicate);