+++ common/src/main/java/me/rufia/fightorflight/net/packet/PokeStaffCmdPacket.java
@@ -9,6 +9,7 @@ public abstract class PokeStaffCmdPacket implements NetworkPacket, CustomPacketP
 
     public PokeStaffCmdPacket(int slot, boolean isFromPokeStaff) {
         this.isFromPokeStaff = isFromPokeStaff;
+        this.slot = slot;
     }
 
     public int getSlot() {