+++ neoforge/src/main/java/me/rufia/fightorflight/forge/FOFForgeClient.java
@@ -2,6 +2,7 @@ package me.rufia.fightorflight.forge;
 
 import me.rufia.fightorflight.CobblemonFightOrFlight;
 import me.rufia.fightorflight.client.keybinds.KeybindFightOrFlight;
+import me.rufia.fightorflight.client.model.PokemonSpikeModel;
 import me.rufia.fightorflight.client.renderer.PokemonArrowRenderer;
 import me.rufia.fightorflight.client.renderer.PokemonBulletRenderer;
 import me.rufia.fightorflight.client.renderer.PokemonSpikeRenderer;
@@ -12,6 +13,7 @@ import net.neoforged.api.distmarker.Dist;
 import net.neoforged.bus.api.SubscribeEvent;
 import net.neoforged.fml.common.EventBusSubscriber;
 import net.neoforged.fml.event.lifecycle.FMLClientSetupEvent;
+import net.neoforged.neoforge.client.event.EntityRenderersEvent;
 import net.neoforged.neoforge.client.event.RegisterKeyMappingsEvent;
 
 @EventBusSubscriber(modid = CobblemonFightOrFlight.MODID, value = Dist.CLIENT, bus = EventBusSubscriber.Bus.MOD)
@@ -27,5 +29,10 @@ public final class FOFForgeClient {
     @SubscribeEvent
     public static void registerKeys(RegisterKeyMappingsEvent event) {
         KeybindFightOrFlight.bindings.forEach(event::register);
+    }
+
+    @SubscribeEvent
+    public static void registerEntityLayer(EntityRenderersEvent.RegisterLayerDefinitions event) {
+        event.registerLayerDefinition(PokemonSpikeModel.LAYER_LOCATION, PokemonSpikeModel::createBodyLayer);
     }
 }